requirements/docs.txt
-r base.txt
-r tests.txt
-r changelog.txt
furo
sphinx
sphinx-copybutton
sphinx-prompt
sphinxcontrib-spelling
sphinxcontrib-towncrier >= 0.2.1a0
autodoc_pydantic
